% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoread.R
\name{autoread}
\alias{autoread}
\title{Auto-read from file}
\usage{
autoread(file, ...)
}
\arguments{
\item{file}{Full path to be read}

\item{...}{Other arguments to be passed to the particular loading function}
}
\value{
The object loaded by the file
}
\description{
Automatically read a file based on extension
}
\seealso{
\code{\link[=get_file_reading_functions]{get_file_reading_functions()}} \code{\link[=autowrite]{autowrite()}}
}
