% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autowrite.R
\name{autowrite}
\alias{autowrite}
\title{Auto-write to file}
\usage{
autowrite(x, file, ...)
}
\arguments{
\item{x}{Object to be saved}

\item{file}{Full path to save the object to}

\item{...}{Other arguments to be passed to the particular saving function}
}
\value{
Invisibly passes TRUE if the file saves successfully
}
\description{
Automatically write an object to a file based on extension
}
\seealso{
\code{\link[=get_file_writing_functions]{get_file_writing_functions()}} \code{\link[=autoread]{autoread()}}
}
