% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autowrite.R
\name{get_file_writing_functions}
\alias{get_file_writing_functions}
\title{Get the list of file writing functions}
\usage{
get_file_writing_functions()
}
\value{
Named list where the names are file extensions, and the values are functions
that read a file. All functions have ... arguments that can be used to extend the
basic function.
}
\description{
Constructs a list of all file-reading functions based on extension
}
\seealso{
\code{\link[=autoread]{autoread()}} \code{\link[=get_file_reading_functions]{get_file_reading_functions()}}
}
