% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{qstop}
\alias{qstop}
\title{Stop without listing the containing function call}
\usage{
qstop(...)
}
\arguments{
\item{...}{Parameters passed to \code{stop()}}
}
\value{
Concisely stops program execution
}
\description{
Stop without listing the containing function call
}
