% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{require_namespace_or_stop}
\alias{require_namespace_or_stop}
\title{Require that a namespace be loaded, or stop execution}
\usage{
require_namespace_or_stop(pkg)
}
\description{
Require that a namespace be loaded, or stop execution
}
