% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{require_namespace_or_stop}
\alias{require_namespace_or_stop}
\title{Require that a namespace be loaded, or stop execution}
\usage{
require_namespace_or_stop(pkg)
}
\arguments{
\item{pkg}{(\code{character(1)}) Package to be loaded}
}
\value{
Silently loads namespace, or stops execution if package cannot be loaded
}
\description{
Require that a namespace be loaded, or stop execution
}
